requirements.txt
markdown==3.4.3
PyYAML==6.0.1
python-frontmatter==1.0.0
rich==13.4.2
requests==2.31.0
resend==0.6.0
groq==0.12.0
